Refactors To into Iterator class
